Extracted hotel data into a separate slice
