+++ src/redux/hooks.ts
@@ -9,5 +9,5 @@ export const useScrollLeft = () => useAppSelector(state => state.main.scrollLeft
 export const useScrollTop = () => useAppSelector(state => state.main.scrollTop);
 export const useStartDate = () => useAppSelector(state => state.main.startDate);
 export const useColumns = () => useAppSelector(state => state.main.columns);
-export const useHotel = () => useAppSelector(state => state.main.hotel);
+export const useHotel = () => useAppSelector(state => state.hotel.data);
 export const useGrabbedTile = () => useAppSelector(state => state.grabbedTile);
